Dev of the C++ version
Non-Functional C++ version of the code.
Initially, only the core functions of the solver will be converted into C++. Those will be tested wrapped in a python code (for graphical handling and solutions verifications). Once validated, it will be considered as the new main.
Then developement will continue to have a 100% based C++ code.